requirements.txt
redis==4.5.1
Django==4.1.5
django-crispy-forms==1.14.0
mysqlclient==2.1.1
django-mysql==4.8.0
django-htmx==1.12.2
coverage==6.5.0
beautifulsoup4==4.11.1
requests==2.28.1
celery==5.2.7
python-dateutil==2.9.0
